move errors into their own submodule
