+++ src/phase0/state_transition/block_processing.rs
@@ -8,15 +8,17 @@ use crate::phase0::operations::{
     Attestation, AttesterSlashing, Deposit, DepositMessage, PendingAttestation, ProposerSlashing,
     SignedVoluntaryExit,
 };
+use crate::phase0::state_transition::errors::{
+    invalid_header_error, invalid_operation_error, Error, InvalidAttestation,
+    InvalidAttesterSlashing, InvalidBeaconBlockHeader, InvalidDeposit, InvalidOperation,
+    InvalidProposerSlashing, InvalidVoluntaryExit,
+};
 use crate::phase0::state_transition::{
     compute_domain, compute_epoch_at_slot, compute_signing_root, get_beacon_committee,
     get_beacon_proposer_index, get_committee_count_per_slot, get_current_epoch, get_domain,
     get_indexed_attestation, get_previous_epoch, get_randao_mix, hash, increase_balance,
-    initiate_validator_exit, invalid_header_error, invalid_operation_error, is_active_validator,
-    is_slashable_attestation_data, is_slashable_validator, is_valid_indexed_attestation,
-    slash_validator, Context, Error, InvalidAttestation, InvalidAttesterSlashing,
-    InvalidBeaconBlockHeader, InvalidDeposit, InvalidOperation, InvalidProposerSlashing,
-    InvalidVoluntaryExit,
+    initiate_validator_exit, is_active_validator, is_slashable_attestation_data,
+    is_slashable_validator, is_valid_indexed_attestation, slash_validator, Context,
 };
 use crate::phase0::validator::Validator;
 use crate::phase0::DEPOSIT_CONTRACT_TREE_DEPTH;
