+++ src/phase0/state_transition/errors.rs
@@ -1,0 +1,198 @@
+use crate::crypto::Signature as BLSSignature;
+use crate::phase0::beacon_block::BeaconBlockHeader;
+use crate::phase0::operations::{AttestationData, Checkpoint};
+use crate::primitives::{Epoch, Root, Slot, ValidatorIndex};
+use ssz_rs::prelude::*;
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error("{0}")]
+    Merkleization(#[from] MerkleizationError),
+    #[error("{0}")]
+    SimpleSerialize(#[from] SimpleSerializeError),
+    #[error("requested element {requested} but collection only has {bound} elements")]
+    OutOfBounds { requested: usize, bound: usize },
+    #[error("invalid signature")]
+    InvalidSignature,
+    #[error("collection cannot be empty")]
+    CollectionCannotBeEmpty,
+    #[error("unable to shuffle")]
+    Shuffle,
+    #[error("slot {requested} is outside of allowed range ({lower_bound}, {upper_bound})")]
+    SlotOutOfRange {
+        requested: Slot,
+        lower_bound: Slot,
+        upper_bound: Slot,
+    },
+    #[error("overflow")]
+    Overflow,
+    #[error("{0}")]
+    InvalidBlock(InvalidBlock),
+    #[error("an invalid transition to a past slot {requested} from slot {current}")]
+    TransitionToPreviousSlot { current: Slot, requested: Slot },
+    #[error("invalid state root")]
+    InvalidStateRoot,
+    #[error(
+    "the requested epoch {requested} is not in the required current epoch {current} or previous epoch {previous}"
+    )]
+    InvalidEpoch {
+        requested: Epoch,
+        previous: Epoch,
+        current: Epoch,
+    },
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidBlock {
+    #[error("invalid beacon block header: {0}")]
+    Header(InvalidBeaconBlockHeader),
+    #[error("invalid operation: {0}")]
+    InvalidOperation(InvalidOperation),
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidOperation {
+    #[error("invalid attestation: {0}")]
+    Attestation(InvalidAttestation),
+    #[error("invalid indexed attestation: {0}")]
+    IndexedAttestation(InvalidIndexedAttestation),
+    #[error("invalid deposit: {0}")]
+    Deposit(InvalidDeposit),
+    #[error("invalid randao (BLS signature): {0:?}")]
+    Randao(BLSSignature),
+    #[error("invalid proposer slashing: {0}")]
+    ProposerSlashing(InvalidProposerSlashing),
+    #[error("invalid attester slashing: {0}")]
+    AttesterSlashing(InvalidAttesterSlashing),
+    #[error("invalid voluntary exit: {0}")]
+    VoluntaryExit(InvalidVoluntaryExit),
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidBeaconBlockHeader {
+    #[error("mismatch between state slot {state_slot} and block slot {block_slot}")]
+    StateSlotMismatch { state_slot: Slot, block_slot: Slot },
+    #[error("mismatch between the block's parent root {expected:?} and the expected parent root {provided:?}")]
+    ParentBlockRootMismatch { expected: Root, provided: Root },
+    #[error("proposer with index {0} is slashed")]
+    ProposerSlashed(ValidatorIndex),
+    #[error("block slot {block_slot} is older than the latest block header slot {latest_block_header_slot}")]
+    OlderThanLatestBlockHeader {
+        block_slot: Slot,
+        latest_block_header_slot: Slot,
+    },
+    #[error("mismatch between the block proposer index {block_proposer_index} and the state proposer index {proposer_index}")]
+    ProposerIndexMismatch {
+        block_proposer_index: ValidatorIndex,
+        proposer_index: ValidatorIndex,
+    },
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidAttestation {
+    #[error("expected length of {expected_length} in bitfield but had length {length}")]
+    Bitfield {
+        expected_length: usize,
+        length: usize,
+    },
+    #[error("invalid target epoch {target}, not current ({current}) or previous epochs")]
+    InvalidTargetEpoch { target: Epoch, current: Epoch },
+    #[error("invalid slot {slot} (in epoch {epoch}) based on target epoch {target}")]
+    InvalidSlot {
+        slot: Slot,
+        epoch: Epoch,
+        target: Epoch,
+    },
+    #[error("attestation at slot {attestation_slot} is not timely for state slot {state_slot}, outside of range [{lower_bound}, {upper_bound}]")]
+    NotTimely {
+        state_slot: Slot,
+        attestation_slot: Slot,
+        lower_bound: Slot,
+        upper_bound: Slot,
+    },
+    #[error("attestation's index {index} exceeds the current committee count {upper_bound}")]
+    InvalidIndex { index: usize, upper_bound: usize },
+    #[error("attestation's source checkpoint {source_checkpoint:?} does not match the expected checkpoint {expected:?} (in epoch {current})")]
+    InvalidSource {
+        expected: Checkpoint,
+        source_checkpoint: Checkpoint,
+        current: Epoch,
+    },
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidIndexedAttestation {
+    #[error("attesting indices are empty")]
+    AttestingIndicesEmpty,
+    #[error("attesting indices are duplicated")]
+    DuplicateIndices(Vec<ValidatorIndex>),
+    #[error("attesting indices are not sorted")]
+    AttestingIndicesNotSorted,
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidDeposit {
+    #[error("expected {expected} deposits but only had {count} deposits")]
+    IncorrectCount { expected: usize, count: usize },
+    #[error("merkle validation failed for tree with depth {depth} and root {root:?} at index {index} for leaf {leaf:?} and branch {branch:?}")]
+    InvalidProof {
+        leaf: Node,
+        branch: Vec<Node>,
+        depth: usize,
+        index: usize,
+        root: Root,
+    },
+    #[error("invalid signature for deposit: {0:?}")]
+    InvalidSignature(BLSSignature),
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidProposerSlashing {
+    #[error("different slots: {0} vs. {1}")]
+    SlotMismatch(Slot, Slot),
+    #[error("different proposers: {0} vs. {1}")]
+    ProposerMismatch(ValidatorIndex, ValidatorIndex),
+    #[error("headers are equal: {0:?}")]
+    HeadersAreEqual(BeaconBlockHeader),
+    #[error("proposer with index {0} is not slashable")]
+    ProposerIsNotSlashable(ValidatorIndex),
+    #[error("header has invalid signature: {0:?}")]
+    InvalidSignature(BLSSignature),
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidAttesterSlashing {
+    #[error("attestation data is not slashable: {0:?} vs. {1:?}")]
+    NotSlashable(Box<AttestationData>, Box<AttestationData>),
+    #[error("no validator was slashed across indices: {0:?}")]
+    NoSlashings(Vec<ValidatorIndex>),
+}
+
+#[derive(Debug, Error)]
+pub enum InvalidVoluntaryExit {
+    #[error("validator is not active in the current epoch {0}")]
+    InactiveValidator(Epoch),
+    #[error("validator {index} already exited in {epoch}")]
+    ValidatorAlreadyExited { index: ValidatorIndex, epoch: Epoch },
+    #[error("exit in epoch {exit_epoch} is not eligible for processing in current epoch {current_epoch}")]
+    EarlyExit {
+        current_epoch: Epoch,
+        exit_epoch: Epoch,
+    },
+    #[error("validator needs to be active for a minimum period of time (from epoch {minimum_time_active}, currently in {current_epoch})")]
+    ValidatoIsNotActiveForLongEnough {
+        current_epoch: Epoch,
+        minimum_time_active: Epoch,
+    },
+    #[error("voluntary exit has invalid signature: {0:?}")]
+    InvalidSignature(BLSSignature),
+}
+
+pub(crate) fn invalid_header_error(error: InvalidBeaconBlockHeader) -> Error {
+    Error::InvalidBlock(InvalidBlock::Header(error))
+}
+
+pub(crate) fn invalid_operation_error(error: InvalidOperation) -> Error {
+    Error::InvalidBlock(InvalidBlock::InvalidOperation(error))
+}
